# swagger.yaml
openapi: 3.0.0
info:
  title: Backend HL Organization API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /admin/transparency/{id}:
    put:
      tags: [Transparency]
      summary: Update transparency file
      requestBody:
        required: true
      responses:
        "200":
          description: File updated
        "400":
          description: Error
    delete:
      tags: [Transparency]
      summary: Delete transparency file
      responses:
        "200":
          description: File deleted
        "400":
          description: Error
  /transparency:
    get:
      tags: [Transparency]
      summary: Get all transparency files
      parameters:
        - in: query
          name: year
          schema:
            type: integer
          required: false
          description: Filter by year
        - in: query
          name: file_type
          schema:
            type: string
          required: false
          description: Filter by file type
      responses:
        "200":
          description: List of transparency files
        "400":
          description: Error
  /transparency/{id}:
    get:
      tags: [Transparency]
      summary: Get transparency file by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Transparency file ID
      responses:
        "200":
          description: Transparency file details
        "404":
          description: File not found
        "400":
          description: Error